requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.0
django-bootstrap5==23.4
django-debug-toolbar==4.3.0
django-filter==23.5
django-rest-auth==0.9.5
django-rosetta==0.10.0
django-summernote==0.8.20.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==22.6.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
polib==1.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
